ts 05

+++ 0x0C-more_malloc_free/101-mul.c
@@ -29,7 +29,7 @@ return (1);
  */
 int main(int argc, char **argv)
 {
-int n1, n2, *mul;
+long int n1, n2, *mul;
 if (argc != 3)
 {
 printf("Error\n");
@@ -40,7 +40,7 @@ if (checkDigit(argv[1]) == 0 || checkDigit(argv[2]) == 0)
 printf("Error\n");
 exit(98);
 }
-mul = malloc(sizeof(int));
+mul = malloc(sizeof(long int));
 if (mul == NULL)
 {
 printf("Error\n");
@@ -49,7 +49,7 @@ exit(98);
 n1 = atoi(argv[1]);
 n2 = atoi(argv[2]);
 *mul = n1 *n2;
-printf("%d\n", *mul);
+printf("%ld\n", *mul);
 free(mul);
 return (0);
 }
